removed vec view

+++ src/gen_vector.c
@@ -12,7 +12,7 @@
 // MACROS
 
 // get ptr to elm at index i
-#define GET_PTR(vec, i) ((vec->data) + ((i) * ((vec)->data_size)))
+#define GET_PTR(vec, i) ((vec->data) + ((u64)(i) * ((vec)->data_size)))
 // get total size in bytes for i elements
 #define GET_SCALED(vec, i) ((i) * ((vec)->data_size))
